\name{centiserve-package}
\alias{centiserve-package}
\alias{centiserve}
\docType{package}
\title{
Functions to find graph centrality indices
}
\description{
Find centrality indices (measures) additional to the igraph package centrality functions. The centiserve is a part of \href{http://www.centiserver.org/}{www.CentiServer.org} project which is a comprehensive centrality measures resource and a web based application for finding graph centrality measures.
}
\details{
\tabular{ll}{
Package: \tab centiserve\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2014-12-30\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Mahdi Jalili \email{m_jalili@farabi.tums.ac.ir}\cr
Adapted algorithms and sources are referenced in function document.

Maintainer:
Mahdi Jalili \email{m_jalili@farabi.tums.ac.ir}
}
\references{
Csardi G, Nepusz T: The igraph software package for complex network research, InterJournal, Complex Systems 1695. 2006. http://igraph.org

Adapted algorithms and sources are referenced in function document.
}